Separate LoggerReporter in two classes
- LoggerReporter only logs
- LoggerTreeReporter extends LoggerReporter and also build up the report tree
